Finish migration from billing_units to API fields
This is now only used for letters and represents the number sent
[^1].  We could use the chargeable_units field, but using "_sent"
is more consistent with the annual attributes [^2].

In fact, chargeable_units isn't actually used anywhere, but I've
kept it in the test data as it is part of the real API and helps
clarify the other values for SMS - free vs. charged.

Note: for SMS I've used an arbitrary "1234" for "chargeable_units"
to indicate it's not used and may be different to the number sent -
for SMS it's related to the number of fragments.

[^1]: https://github.com/alphagov/notifications-api/blob/bb62d22f2555db192ab9de66440409158524d790/app/dao/fact_billing_dao.py#L339
[^2]: https://github.com/alphagov/notifications-admin/blob/3a1ac189ff6800d4789b22e4763e5550719e8802/app/main/views/dashboard.py#L339