Stop using `_external=True` in tests
It looks like, by default, Flask no longer makes full URLs, for example
`https://example.com/path`. Instead it does `/path`. This will still
work fine, and if anything is better because it reduces the number of
bytes of HTML we are sending.

It won’t mean that requests go over `http` instead of `https` without
the protocol because we set the appropriate HSTS header here:
https://github.com/alphagov/notifications-aws/blob/0c57da7781260c56575b796062795a9d8e1f8c87/ansible/roles/paas-proxy/templates/admin.conf.j2#L11

This commit changes all our tests to reflect that URLs no longer have
the protocol and domain in them. `_external=True` is Flask’s way of
saying whether a URL should be generated with the domain and protocol
(`True`) or without it (`False`).

Again, I can’t find the changelog or diff where this was introuduced,
but if you’d like to go spelunking then here’s a starting point:
https://github.com/pallets/flask/blob/50374e3cfe815acd24fd390c861561a0dd17bfbb/src/flask/helpers.py#L192