Convert frontend display to be just UTC
This changeset converts the display of dates and times to be just UTC to match the recent changes in the backend.  This unwinds a bit of work that was done previously and allows us to start with a clean slate in how we want to approach displaying dates and times going forward.

There are still a few remaining TODOs:

- Add information about how different timezones map to UTC
- Fix all of the broken tests
- Check for any other places that still need to be changed

Signed-off-by: Carlo Costino <carlo.costino@gsa.gov>