Made several adjustments for session timer tests
- Turned the timeoutPopup module into a proper module that is exported
- Pulled out the setInterval inner function to be its own method
- Adjusted the session timer tests to have additional setup and teardown pieces
- Added a few tests

There may be some additional adjustments needed, e.g., we may not have to expose all of the methods, but then we have to figure out how to make them accessible in the tests another way.

h/t @A-Shumway42 for pairing with me to review all of these changes and approach!

Signed-off-by: Carlo Costino <carlo.costino@gsa.gov>