Adjust E2E test fixtures
This changeset adjusts the scope of the end-to-end (E2E) test fixtures so that they behave as intended and we have truly isolated tests.  We may adjust these again in the future depending on our needs, but for right now this will hopefully keep things simple for us as we start to build out an E2E test library.

Thank you, @terrazoon, for flagging this!

Signed-off-by: Carlo Costino <carlo.costino@gsa.gov>