Restore original sign-in form
This changeset restores the original sign-in form that we hid so that folks can still sign into Notify.gov and have full access to their accounts.

There is currently an issue that we are investigating where some users do not have full account access when signing in via Login.gov, so we are providing this as a work around until we can resolve the full issue.

Signed-off-by: Carlo Costino <carlo.costino@gsa.gov>