Remove E2E environment variables from the manifest
This changeset removes the E2E environment variables from the admin manifest file.  While these might be needed on the API side to ensure that they are exposed and able to be used in the staging environment, we do not need them to be enabled for the admin app; they should only be present in the GitHub Actions that are run in a PR, and the .github/workflows/checks.yml file takes care of that and injects them directly into the container environment that the end-to-end tests run in.

Signed-off-by: Carlo Costino <carlo.costino@gsa.gov>